budget/requirements.txt
djangorestframework==3.13.1
Django==3.2.15
django-filter==22.1

